Allow to specify a transformation on those transitions going to transient states.
you can now optionally specify a transformation on a transition going to
a transient state. Since a transient state will simply have it's
"onEvent" executed right away there is sometimes a need to allow to
transform the event to something else.

As a nice side effect of transient states and transformations (although
transformations just makes it more useful) is that you can build up
entire decision trees on transient states only and using branches
(transitions with guards) and transformations.

Also:

* extracted out interfaces for all various builders.
* Upgraded to latest jackson due to security vulnerability