Enhanced the Schedular to allow for a producer. Changed the FSM Context
* The Actor Schedular now allows to pass in a Producer<T> to produce the
msg to send. This then aligns better with the FSM Schedular.
* The FSM Context now follows the pattern of the Actor in that the FSM
Schedular that is accessible on the FSM Context, is now stored in a
ThreadLocal. Currently though, since the Context is an interface there
is no way to prevent the user from overriding this behavior. However,
for now, it'll have to do.