Fixed shutdown bug
Shutdown wasn't properly implemented, now it is :-).

Note: this current shutdown version is quite destructive in that it
doesn't allow all potential messages on the internal job queue to finish
first. That means that actors may not process all messages that are in
the queue. An extension to the current one would perhaps to implement a
`shutdown` and a `shutdownNow`, which is a common approach elsewhere.

For now, this current version essentially works as a `shutdownNow`